Conflicts:
	PirateGame/.vs/PirateGame/v14/.suo
	PirateGame/PirateGame/Game1.cs
	PirateGame/PirateGame/PirateGame.csproj

	PirateGame/PirateGame/obj/x86/Debug/DesignTimeResolveAssemblyReferencesInput.cache
	PirateGame/PirateGame/obj/x86/Debug/PirateGame.application

	PirateGame/PirateGame/obj/x86/Debug/PirateGame.csproj.FileListAbsolute.txt
	PirateGame/PirateGame/obj/x86/Debug/PirateGame.exe
	PirateGame/PirateGame/obj/x86/Debug/PirateGame.exe.manifest
	PirateGame/PirateGame/obj/x86/Debug/PirateGame.pdb

+++ PirateGame/PirateGame/Ship.cs
@@ -39,7 +39,7 @@ namespace PirateGame
             return PowerLevel;
         }
 
-        public float getGold()
+        public int getGold()
         {
             return gold;
         }
